修改时间时区同步模块时间戳获取机制

+++ services/time_manager/include/sntp_client.h
@@ -144,7 +144,7 @@ private:
      * @param _ntpTs the NTP timestamp to be converted
      * Returns the milliseconds
      */
-    uint64_t ConvertNtpToStamp(uint64_t _ntpTs);
+    int64_t ConvertNtpToStamp(uint64_t _ntpTs);
 
     /**
      * This function converts the NTP time to local time
@@ -154,8 +154,8 @@ private:
      */
     uint64_t ConvertNtpToDate(uint64_t _ntpTs, struct date_structure* _outDataTs);
     
-    int m_clockOffset;
-    int64_t m_originateTimestamp;
+    int64_t m_clockOffset;
+    uint64_t m_originateTimestamp;
     int64_t mNtpTime;
     int64_t mNtpTimeReference;
     int64_t mRoundTripTime;
